Avoid an overflow when initializing an __mpz_struct for an IntInf result.
The _mp_alloc field of an __mpz_struct is declared as an int.  On a
64-bit architecture, the computation of the number of available limbs
from the space available in the heap occurs at 64-bits.  With
>8Gb heap and sizeof(mp_limb_t) == 4 or >16Gb heap and
sizeof(mp_limb_t) == 8, the result will overflow when being assigned
to an int.


SVN r6527

+++ runtime/gc/int-inf.h
@@ -1,4 +1,4 @@
-/* Copyright (C) 1999-2007 Henry Cejtin, Matthew Fluet, Suresh
+/* Copyright (C) 1999-2008 Henry Cejtin, Matthew Fluet, Suresh
  *    Jagannathan, and Stephen Weeks.
  * Copyright (C) 1997-2000 NEC Research Institute.
  *
@@ -54,7 +54,7 @@ COMPILE_TIME_ASSERT(sizeof_mp_limb_t__compat__sizeof_objptr,
                     (sizeof(objptr) % sizeof(mp_limb_t) == 0));
 #define LIMBS_PER_OBJPTR ( \
         sizeof(mp_limb_t) >= sizeof(objptr) ? \
-        1 : sizeof(objptr) / sizeof(mp_limb_t))
+        1 : (int)(sizeof(objptr) / sizeof(mp_limb_t)))
 
 void initIntInf (GC_state s);
 static inline void fillIntInfArg (GC_state s, objptr arg, __mpz_struct *res,
